data_quality/requirements.txt
pyspark==3.5.0
delta-spark==3.0.0
wheel==0.42.0
setuptools==69.1.0
autopep8==2.0.4